Finishing the app

+++ pages/api/products/[id].js
@@ -4,8 +4,11 @@ import Product from "../../../models/Product";
 export default async function handler(req, res) {
   const {
     method,
+    cookies,
     query: { id },
   } = req;
+
+  const token = cookies.token;
 
   await dbConnect();
 
@@ -19,9 +22,23 @@ export default async function handler(req, res) {
   }
 
   if (method === "PUT") {
+    if (!token || token !== process.env.TOKEN) {
+      return res.status(401).json("Not Authenticated !!");
+    }
+    try {
+      const product = await Product.findByIdAndUpdate(id, req.body, {
+        new: true,
+      });
+      res.status(201).json(product);
+    } catch (error) {
+      res.status(500).json(error);
+    }
   }
 
   if (method === "DELETE") {
+    if (!token || token !== process.env.TOKEN) {
+      return res.status(401).json("Not Authenticated !!");
+    }
     try {
       await Product.findByIdAndDelete(id);
       res.status(200).json("The Product Has Been Deleted");
